Fix missing dll error
A missing dll was being picked up at run time but not compile time. Lots
of logging was added while trying to fix that and I've left it in as
it's useful for me and anyone else using DataShare.

Also removed projects which were no longer used and changed some of the
build configuration information.